+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 * Support --locale command line option
 * Support --preferences command line option
+* Support --only command line option
 
 # 0.1.0